Fixes to variogram fit for some edge cases
Set min valid range for power variograms to 1.8. This value is theoretically 2, but gsmap produces all nan interpolations for power variograms with range > 1.8. Will try reporting the bug to intamap.
Fix error handling in afvGLS function. Before if a variogram model failed to be fit via GLS but was fit via the traditional method, the traditional estimate would be discarded, now we keep it and use RMSE wrt variogram cloud to determine which of the fit variograms is the best (this is what we did before).
Add some more logging and error handling